Refactor monkey patch "PatchCopyContainerVerifyObjectPaste".
